requirements.txt
Django==4.2.7
django-cors-headers==4.7.0
django-filter==25.1
django-import-export==4.3.7
django-rest-framework==0.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
mysqlclient==2.2.7
requests